refactor(renderer): worker-safe white-pixel helper as Renderer static
Two PR review threads pointed out that the multi-material work added
bare `document.createElement("canvas")` calls in Mesh and the Canvas
renderer — these throw in OffscreenCanvas / worker contexts where
`document` is undefined.

Introduce `Renderer.getWhitePixel()` — a lazily-created, shared 1×1
white canvas using `OffscreenCanvas` where supported and falling back
to `document.createElement` otherwise. Static (not instance-bound) so
it's reachable from `Mesh` construction before any active renderer
exists.

Mesh's `isMultiMaterial` Kd-only path now uses
`Renderer.getWhitePixel()` instead of its own local
`getOrCreateWhitePixel()`; CanvasRenderer's `_meshColorCanvas`
scratch sampler now goes through the same OffscreenCanvas-aware
allocation pattern.

Kept inline (not via `CanvasRenderTarget`) — that abstraction is
sized for full-blown render targets with their own GL context, way
heavier than what a 1x1 fallback needs. Also kept off the `video`
namespace — `video.createCanvas` is on a deprecation path; canvas
allocation is renderer-side concern now.
